+++ CHANGELOG.md
@@ -27,10 +27,8 @@
       * Github: [PR-561](https://github.com/pegasystems/angular-sdk-components/pull/561)
 *   **Fixed label display in Details templates.**
       * Github: [PR-568](https://github.com/pegasystems/angular-sdk-components/pull/568)
-*   **Fixed the issue where the DataReference value was displayed as a SemanticLink in the CaseSummary view.**
-      * Github: [PR-572](https://github.com/pegasystems/angular-sdk-components/pull/572)
 *   **Displays the configured custom label for the Add button.**
-    * Github: [PR-574](https://github.com/pegasystems/angular-sdk-components/pull/574)
+      * Github: [PR-574](https://github.com/pegasystems/angular-sdk-components/pull/574)
 
 # [25.1.13](https://github.com/pegasystems/angular-sdk/tree/release/25.1.13) - Released: 12/06/2026