+++ CHANGELOG.md
@@ -13,6 +13,8 @@
     * Github: [PR-545](https://github.com/pegasystems/angular-sdk-components/pull/545)
 *   **Fixed DataReference field value rendering in the Details template.**
     * Github: [PR-562](https://github.com/pegasystems/angular-sdk-components/pull/562)
+*   **Added left and right alignment support for vertical multi-step assignment navigation.**
+    * Github: [PR-576](https://github.com/pegasystems/angular-sdk-components/pull/576)
 
 ### **Bug fixes**
 *   **Fixed the issue where views are not rendering in Details Template.**