+++ CHANGELOG.md
@@ -11,7 +11,7 @@
     * Github: [PR-553](https://github.com/pegasystems/angular-sdk-components/pull/553)
 *   **Replaced the TinyMCE dependency with Tiptap for rich text editing capabilities.**
     * Github: [PR-545](https://github.com/pegasystems/angular-sdk-components/pull/545)
-*   **Fixed DataReference field value rendering in the Details template.**
+*   **Added Support for DataReference field value rendering in the Details template.**
     * Github: [PR-562](https://github.com/pegasystems/angular-sdk-components/pull/562)
 *   **DataReference as Autocomplete supports Secondary Text.**
     * Github: [PR-569](https://github.com/pegasystems/angular-sdk-components/pull/569)
@@ -29,6 +29,8 @@
       * Github: [PR-561](https://github.com/pegasystems/angular-sdk-components/pull/561)
 *   **Fixed label display in Details templates.**
       * Github: [PR-568](https://github.com/pegasystems/angular-sdk-components/pull/568)
+*   **Fixed the issue where the DataReference value was displayed as a SemanticLink in the CaseSummary view.**
+      * Github: [PR-572](https://github.com/pegasystems/angular-sdk-components/pull/572)
 *   **Displays the configured custom label for the Add button.**
       * Github: [PR-574](https://github.com/pegasystems/angular-sdk-components/pull/574)